Refactoring
Refactored certain static methods out of PlantedCrop into GridUtilities and PlantableArea
