+++ Assets/Scripts/PlantableArea.cs
@@ -23,6 +23,25 @@ public class PlantableArea : MonoBehaviour
         return plantableBoundary.Contains(point);
     }
 
+    /*
+    Checks if a provided position is in a plantable area, i.e. a GameObject with the PlantableArea tag that also has a PlantableArea script attached
+    @return True if the provided coordinate was in any plantable area, false if not
+    */
+    public static bool CheckInPlantableArea(Vector3 position) {
+        GameObject[] areas = GameObject.FindGameObjectsWithTag("PlantableArea");
+        if(areas.Length == 0) //Return if no areas were found with the tag
+            return false;
+
+        foreach(GameObject obj in areas) {
+            PlantableArea area = obj.GetComponent<PlantableArea>();
+            if((area != null) & area.ContainsPoint(position)) {
+                return true;
+            }
+        }
+
+        return false;
+    }
+
     //Draw debug outlines
     void OnDrawGizmos() {
         Gizmos.color = Color.yellow;
