+++ Assets/Scripts/Trap.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 541d5268abba879478084412069e83f2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
